+++ src/main/java/wooteco/subway/dto/path/PathResponse.java
@@ -6,10 +6,10 @@ import wooteco.subway.dto.station.StationResponse;
 public class PathResponse {
 
     private final List<StationResponse> stations;
-    private final int distance;
+    private final long distance;
     private final int fare;
 
-    public PathResponse(List<StationResponse> stations, int distance, int fare) {
+    public PathResponse(List<StationResponse> stations, long distance, int fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -19,7 +19,7 @@ public class PathResponse {
         return stations;
     }
 
-    public int getDistance() {
+    public long getDistance() {
         return distance;
     }
